Update devel/apache-maven to 2.2.0
Maven 2.2.0 contains a few important changes that justify the version
upgrade, instead of simply naming it 2.1.1. First, the Java requirement
for Maven 2.2.0 has been upgraded to 1.5 or later. This upgrade was
planned for 2.1.0, but that release still contained binaries that were
compatible with JDK 1.4. In addition, due to some serious flaws in the
version-expression POM transformation included in 2.1.0, this feature
has been removed for the time being. Finally, some new default execution
IDs have been added to Maven to enable the separation of configuration
for plugins bound by the default lifecycle mappings, and for those
invoked directly from the command line.

Changes that may affect existing builds

* MNG-4143 - Starting in 2.2.0, Maven will run only on Java 1.5 and later.
  You can still build projects for JDK1.4 and earlier using the approach
  documented in the Guide to Building JDK 1.4 Projects on JDK 1.5.
* MNG-3401 - Executions with an id equal to default-phase (where phase is
  a valid lifecycle phase) may have unexpected results as it will be merged
  into the default lifecycle.
* MNG-4140/4179 - Version-expression resolution during installation and
  deployment has been removed, returning to Maven 2.0.x behaviour.