Update to 2.8.0:
We are pleased to present this major release of the xf86-video-intel
driver. This release is very similar to the 2.7.99.902 release
candidate but includes a couple of additional bug fixes, (for bugs
#22760 and @22483).

Compared to the 2.7 releases this driver contains several new features
and many bug fixes. The driver now depends on X server 1.6 or later,
and eliminates several obsolete code paths, (XAA and EXA removed in
favor of UXA, DRI1 support eliminated). The driver certainly will work
best with an i915 module from a recent kernel (2.6.31) and with
kernel-modesetting (KMS, specified by loading i915 with the option
"modeset=1").