Update to 1.0, using pkgsrc-wip/kmymoney2-stable packaged by Gary Duzan.
Since our latest stable release, 0.8.9, a lot of effort has been
put by the developers and the community to add new features and
test them to make this a rock-solid release. Over 2 years of
development have seen the addition of budgets, a forecast feature,
many new reports, report charts, a complete redesign of the import
feature, which allows for a much easier migration from other
application and a swifter synchronization with online banking. The
support of PGP encryption for the KMyMoney files has been improved
too, including the option to have multiple keys for a single file,
so no one can access your financial your financial records. The
summary view has been revamped to show more and more useful
information, allowing you to have an overview of your finance at
a glance. Also, there are now translations for 22 languages, though
not all of them are as complete as we would like, and we have users
wherever KDE3 is installed. That results not only in a greater
quality application, but also in one that can be customized to fit
the needs of a wide range of users. Additionally, if you use either
Oxygen or Tango icons, you can enjoy the exclusive sets of icons
we have prepared for this release, because finance can be good
looking too! In between all that work, we have fixed a lot of bugs
and little annoyances to make this the best KMyMoney release ever!

Into the future, the effort to port the application to KDE4 is
already underway. Following our usual conservative approach, the
first release for KDE4 will be very similar to this one, featurewise.
Subsequent releases will take full advantage of the new platform.
Also, the coordination work with other KDE4 financial applications
under the umbrella of the nascient KDE Finance group opens the door
to other exciting possibilities.