import the frama-c source code analysis tool
Frama-C is a suite of tools dedicated to the analysis of the source code of
software written in C.

Frama-C gathers several static analysis techniques in a single collaborative
framework. The collaborative approach of Frama-C allows static analyzers to
build upon the results already computed by other analyzers in the framework.
Thanks to this approach, Frama-C provides sophisticated tools, such as a slicer
and dependency analysis.